Finished refactoring and modified db references
